+++ RssWatcher.h
@@ -8,7 +8,7 @@ public:
 	CRssWatcher();
 
 	DWORD Run();
-	void ProcessFeed(__int64 id, const CString& sUrl);
+	void ProcessFeed(int id, const CString& sUrl);
 	CEvent shutDown; // CNewzflow will set this event to indicate the dir watcher should shut down
 
 protected:
